pyproject.toml
[tool.poetry]
name = "multimodal-rag"
version = "0.1.0"
description = "Building a highly-scalable RAG system on video"
authors = ["MarkusKhoa <dangkhoa20006@gmaill.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "3.12.3"
moviepy = "^1.0.3"
pytube = "^15.0.0"
yt-dlp = "^2025.1.26"
loguru = "^0.7.2"
fastapi = "^0.110.0"
openai-whisper = "^20240930"
langchain = "^0.3.15"
pydantic = "^2.10.5"
pydub = "^0.25.1"
secretstorage = "^3.3.3"
openai = "^1.61.1"
python-dotenv = "^1.0.1"
faster-whisper = "^1.1.1"
vllm = "^0.7.2"
tritonclient = "^2.54.0"
pytest = "^8.3.4"
prometheus-client = "^0.21.1"
ray = "^2.42.1"
langchain-groq = "^0.2.4"
weaviate-client = "^4.10.4"
tiktoken = "^0.9.0"
pypdf = "^5.3.0"
rapidocr-onnxruntime = "^1.4.4"
pyvi = "^0.1.1"
langchain-community = "^0.3.17"
sentence-transformers = "^3.4.1"
langchain-huggingface = "^0.1.2"
langchain-core = "^0.3.35"
langchain-weaviate = "^0.0.4"
langchain-text-splitters = "^0.3.6"
langchain-openai = "^0.3.6"
langchain-milvus = "^0.1.8"
opentelemetry-distro = "^0.52b1"
google-cloud-storage = "^3.1.0"
kubernetes = "^32.0.1"
uvicorn = "^0.34.2"
python-multipart = "^0.0.20"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.10.0"
isort = "^5.13.2"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
